requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.6
django-breadcrumbs==1.1.3
django-ckeditor==6.7.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
filelock==3.15.4
pillow==10.4.0
platformdirs==4.2.2
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.3
